refactor(pathToPosix): move to utl folder
document why it doesn't take win32 exceptions into account
