+++ src/extract/extract.js
@@ -5,7 +5,7 @@ const path                        = require('path').posix;
 const acorn                       = require('acorn');
 const acorn_loose                 = require('acorn/dist/acorn_loose');
 const _                           = require('lodash');
-const pathToPosix                 = require('./resolve/pathToPosix');
+const pathToPosix                 = require('../utl/pathToPosix');
 const resolve                     = require('./resolve');
 const transpile                   = require('./transpile');
 const ignore                      = require('./ignore');
